+++ api_yamdb/api/mixins.py
@@ -1,0 +1,8 @@
+from rest_framework import mixins, viewsets
+
+
+class ListCreateDeleteViewSet(mixins.ListModelMixin,
+                              mixins.CreateModelMixin,
+                              mixins.DestroyModelMixin,
+                              viewsets.GenericViewSet):
+    pass
