--- api_yamdb/api/permission.py
@@ -1,18 +1,0 @@
-from rest_framework import permissions
-
-
-class IsAdminModeratorAuthorOrReadOnly(permissions.BasePermission):
-
-    def has_permission(self, request, view):
-        if request.method in permissions.SAFE_METHODS:
-            return True
-        return request.user.is_authenticated
-
-    def has_object_permission(self, request, view, obj):
-        if request.method in permissions.SAFE_METHODS:
-            return True
-        return (
-            request.user.is_admin
-            or request.user.is_moderator
-            or obj.author == request.user
-        )
