+++ api_yamdb/api/urls.py
@@ -1,6 +1,8 @@
 from django.urls import include, path
 from rest_framework.routers import DefaultRouter
-from .views import UserViewSet, signup, token, ReviewViewSet, CategoryViewSet, GenreViewSet, TitleViewSet
+from .views import (UserViewSet, signup, token, ReviewViewSet,
+                    CategoryViewSet, GenreViewSet, TitleViewSet,
+                    CommentViewSet)
 
 
 router = DefaultRouter()
@@ -14,6 +16,11 @@ router.register(
 )
 router.register(
     'titles', TitleViewSet, basename='titles'
+)
+router.register(
+    r'titles/(?P<title_id>\d+)/reviews/(?P<review_id>\d+)/comments',
+    CommentViewSet, basename='comments'
+)
 
 authurls = [
     path("signup/", signup, name="signup"),
@@ -23,4 +30,4 @@ authurls = [
 urlpatterns = [
     path("v1/", include(router.urls)),
     path("v1/auth/", include(authurls)),
-
+]
