+++ api_yamdb/api_yamdb/settings.py
@@ -13,8 +13,8 @@ DEBUG = True
 ALLOWED_HOSTS = ["*"]
 
 
+#Application definition
 
-# Application definition
 
 INSTALLED_APPS = [
     "django.contrib.admin",
