+++ api_yamdb/reviews/models.py
@@ -1,4 +1,3 @@
-from django.contrib.auth import get_user_model
 from django.db import models
 from django.core.validators import (
     MinLengthValidator,
@@ -6,8 +5,6 @@ from django.core.validators import (
 )
 from django.contrib.auth.models import AbstractUser
 from django.contrib.auth.validators import ASCIIUsernameValidator
-
-User = get_user_model()
 
 
 class Role:
@@ -49,45 +46,52 @@ class User(AbstractUser):
     REQUIRED_FIELDS = ["email"]
 
 
-    
-class Comment(models.Model):
-    review = models.ForeignKey(
-        Review, on_delete=models.CASCADE,
-        related_name='comments',
-        verbose_name='Отзыв'
+class Categories(models.Model):
+    name = models.CharField(max_length=256, unique=True)
+    slug = models.SlugField(max_length=50, unique=True)
+
+    def __str__(self):
+        return self.name
+
+
+class Genres(models.Model):
+    name = models.CharField(max_length=256, unique=True)
+    slug = models.SlugField(max_length=50, unique=True)
+
+    def __str__(self):
+        return self.name
+
+
+class Titles(models.Model):
+    name = models.CharField(max_length=256)
+    year = models.PositiveSmallIntegerField()
+    description = models.TextField(blank=True, null=True)
+    genre = models.ManyToManyField(
+        Genres,
+        related_name='titles',
+        through='TitlesGenres',
     )
-    text = models.TextField(verbose_name='Текст')
-    author = models.ForeignKey(
-        User, on_delete=models.CASCADE,
-        related_name='comments',
-        verbose_name='Автор'
-    )
-    pub_date = models.DateTimeField(
-        auto_now_add=True,
-        verbose_name='Дата публикации'
+    category = models.ForeignKey(
+        Categories,
+        on_delete=models.SET_NULL,
+        related_name='titles',
+        null=True
     )
 
     class Meta:
-        constraints = [
-            models.UniqueConstraint(
-                fields=['author', 'title'],
-                name='unique review'
-            )
-        ]
-        verbose_name = 'Комментарий'
-        verbose_name_plural = 'Комментарии'
-        ordering = ('-pub_date',)
+        verbose_name = 'Произведение'
+        verbose_name_plural = 'Произведения'
 
     def __str__(self):
-        return self.text
-      
-      
+        return self.name
+
+
 class Review(models.Model):
     title = models.ForeignKey(
-        Title,
+        Titles,
         on_delete=models.CASCADE,
         related_name='reviews',
-        verbose_name='Произведение'
+        verbose_name='произведение'
     )
     text = models.TextField()
     author = models.ForeignKey(
@@ -117,46 +121,36 @@ class Review(models.Model):
         ]
         verbose_name = 'Отзыв'
         verbose_name_plural = 'Отзывы'
-        ordering = ('-pub_date')
+        ordering = ('-pub_date', )
 
     def __str__(self) -> str:
         return self.text
 
-      
-class Categories(models.Model):
-    name = models.CharField(max_length=256, unique=True)
-    slug = models.SlugField(max_length=50, unique=True)
 
-    def __str__(self):
-        return self.name
-
-
-class Genres(models.Model):
-    name = models.CharField(max_length=256, unique=True)
-    slug = models.SlugField(max_length=50, unique=True)
-
-    def __str__(self):
-        return self.name
-
-
-class Titles(models.Model):
-    name = models.CharField(max_length=256)
-    year = models.PositiveSmallIntegerField()
-    description = models.TextField(blank=True, null=True)
-    genre = models.ManyToManyField(
-        Genres,
-        related_name='titles',
-        through='TitlesGenres',
-     )
-    category = models.ForeignKey(
-        Categories,
-        on_delete=models.SET_NULL,
-        related_name='titles',
-        null=True
+class Comment(models.Model):
+    review = models.ForeignKey(
+        Review, on_delete=models.CASCADE,
+        related_name='comments',
+        verbose_name='Отзыв'
+    )
+    text = models.TextField(verbose_name='Текст')
+    author = models.ForeignKey(
+        User, on_delete=models.CASCADE,
+        related_name='comments',
+        verbose_name='Автор'
+    )
+    pub_date = models.DateTimeField(
+        auto_now_add=True,
+        verbose_name='Дата публикации'
     )
 
+    class Meta:
+        verbose_name = 'Комментарий'
+        verbose_name_plural = 'Комментарии'
+        ordering = ('-pub_date', )
+
     def __str__(self):
-        return self.name
+        return self.text
 
 
 class TitlesGenres(models.Model):
